+++ StateMenu.cpp
@@ -1,3 +1,6 @@
+/// Author: Matthew Lowe
+/// Copyright (c) 2018
+
 #include "StateMenu.hpp"
 
 using namespace tle;
@@ -24,7 +27,7 @@ void CStateMenu::Init() {
 	// Align the menu items
 	for (auto &item : mItems)
 	{
-		item.second.mPos = Vec2<>(mSettings.GetMenuPos().x, i * mSettings.GetMenuSpacing() + mSettings.GetMenuPos().y);
+		item.second.mPos = Vec2<>((int)mSettings.GetMenuPos().x, i * mSettings.GetMenuSpacing() + (int)mSettings.GetMenuPos().y);
 		++i;
 	}
